+++ src/main/java/sp/data/dao/interfaces/OrderDao.java
@@ -4,8 +4,8 @@ package sp.data.dao.interfaces;
 import sp.data.dao.generic.GenericDao;
 import sp.data.entities.Order;
 
-public interface OrderDao extends GenericDao<Order, Integer>{
+public interface OrderDao extends GenericDao<Order,Long> {
 
-    Order getByIdWithAllChildren(int id);
+    Order getByIdWithAllChildren(Long id);
 
 }
