+++ src/main/java/sp/data/dao/interfaces/OrderPositionDao.java
@@ -4,7 +4,7 @@ package sp.data.dao.interfaces;
 import sp.data.dao.generic.GenericDao;
 import sp.data.entities.OrderPosition;
 
-public interface OrderPositionDao extends GenericDao<OrderPosition, Integer>{
+public interface OrderPositionDao extends GenericDao<OrderPosition,Long> {
 	
 	//TODO: add some specific methods
