+++ src/main/java/sp/data/services/interfaces/OrderPositionService.java
@@ -4,7 +4,7 @@ package sp.data.services.interfaces;
 import sp.data.entities.OrderPosition;
 import sp.data.services.generic.GenericService;
 
-public interface OrderPositionService extends GenericService<OrderPosition, Integer> {
+public interface OrderPositionService extends GenericService<OrderPosition, Long> {
 
 	
 }
