+++ src/main/java/sp/data/services/interfaces/OrderService.java
@@ -7,8 +7,8 @@ import sp.data.services.generic.GenericService;
 import java.math.BigDecimal;
 import java.util.List;
 
-public interface OrderService extends GenericService<Order, Integer> {
+public interface OrderService extends GenericService<Order, Long> {
 
-    Order getByIdWithAllChildren(int id);
+    Order getByIdWithAllChildren(Long id);
 	
 }
